// data/github/current/ui/packages/memex/src/client/api/memex-items/paginated-views.ts
export type PageInfo = {
  hasNextPage: boolean
  hasPreviousPage: boolean
  endCursor?: string
  startCursor?: string
}

export type GroupedMemexItems<TMemexItemType> = {
  groupId: string
  secondaryGroupId?: string
  nodes: Array<TMemexItemType>
  pageInfo: PageInfo
}